+++ src/cli.ts
@@ -7,7 +7,6 @@ import * as opentype from "opentype.js";
 import PngQuant from "pngquant";
 import { generateBitmap } from "./generateBitmap";
 import type { BmpfontGeneratorCliConfig, FontRenderingOptions, GlyphSourceTable, SizeOptions } from "./type";
-import { charsToGlyphList } from "./util";
 
 export async function run(argv: string[]): Promise<void> {
 	const config = parseArguments(argv);
@@ -33,9 +32,6 @@ async function app(param: BmpfontGeneratorCliConfig): Promise<void> {
 	};
 
 	const chars: string[] = param.chars.split("");
-	// if (param.missingGlyph) chars.push(param.missingGlyph);
-
-	// const charAndMissingGlyph: (string | { key: string, src: string | canvas.Image })[] = chars;
 	const sourceTable: GlyphSourceTable<string | canvas.Image> = chars.reduce((table, ch) => {
 		table[ch.charCodeAt(0)] = ch;
 		return table;
